Fixed configurations for all serviceFixed configurations for all servicess
