+++ commerce-service/pom.xml
@@ -96,7 +96,7 @@
                         <image>ecommerce/${project.artifactId}:${project.version}</image>
                     </to>
                     <container>
-                        <mainClass>com.ecommerce.license.LicenseServiceApplication</mainClass>
+                        <mainClass>com.ecommerce.commerce.CommerceServiceApplication</mainClass>
                         <ports>
                             <port>8080</port>
                         </ports>
