+++ commerce-service/src/main/java/com/ecommerce/commerce/CommerceServiceApplication.java
@@ -1,4 +1,4 @@
-package com.ecommerce;
+package com.ecommerce.commerce;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.cloud.context.config.annotation.RefreshScope;
 
 @SpringBootApplication
 @RefreshScope
-public class CustomerServiceApplication {
+public class CommerceServiceApplication {
     public static void main(String[] args) {
-        SpringApplication.run(CustomerServiceApplication.class, args);
+        SpringApplication.run(CommerceServiceApplication.class, args);
     }
 }
