+++ commerce-service/src/main/resources/application.yml
@@ -1,0 +1,10 @@
+spring:
+  application:
+    name: commerce-service
+  profiles:
+    active: ${SPRING_PROFILES_ACTIVE}
+  cloud:
+    config:
+      uri: ${SPRING_CLOUD_CONFIG_URI}
+  config:
+    import: "configserver:${SPRING_CLOUD_CONFIG_URI}"
