+++ commerce-service/src/test/java/com/ecommerce/commerce/CommerceServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.ecommerce.commerce;
+
+import org.junit.jupiter.api.Disabled;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class CommerceServiceApplicationTests {
+
+    @Disabled
+    void contextLoads() {
+    }
+
+}
