+++ compose.yaml
@@ -1,9 +1,70 @@
 services:
-  postgres:
+#  commerce-db:
+#    image: 'postgres:latest'
+#    environment:
+#      - 'POSTGRES_DB=commerce_db'
+#      - 'POSTGRES_PASSWORD=commerce'
+#      - 'POSTGRES_USER=commerce'
+#    ports:
+#      - '5432:5432'
+#    healthcheck:
+#      test: ["CMD-SHELL", "pg_isready -U commerce -d commerce_db"]
+#      interval: 10s
+#      timeout: 5s
+#      retries: 5
+
+  customer-db:
     image: 'postgres:latest'
     environment:
-      - 'POSTGRES_DB=mydatabase'
-      - 'POSTGRES_PASSWORD=secret'
-      - 'POSTGRES_USER=myuser'
+      - 'POSTGRES_DB=consumer_db'
+      - 'POSTGRES_PASSWORD=consumer'
+      - 'POSTGRES_USER=consumer'
     ports:
-      - '5432'
+      - '5433:5433'
+    healthcheck:
+      test: ["CMD-SHELL", "pg_isready -U consumer -d consumer_db"]
+      interval: 10s
+      timeout: 5s
+      retries: 5
+
+  configserver:
+    image: ecommerce/configserver:0.0.1-SNAPSHOT
+    ports:
+      - "8089:8089"
+    environment:
+      SERVER_PORT: "8089"
+      ENCRYPT_KEY: "fje83Ki8403Iod87dne7Yjsl3THueh48jfuO9j4U2hf64Lo"
+
+#  commerce-service:
+#    image: ecommerce/commerce-service:0.0.1-SNAPSHOT
+#    environment:
+#      SPRING_PROFILES_ACTIVE: "dev"
+#      SPRING_CLOUD_CONFIG_URI: "http://configserver:8089"
+#      CONFIGSERVER_PORT: "8089"
+#      DATABASESERVER_PORT: "5432"
+#      ENCRYPT_KEY: "IMSYMMETRIC"
+#    depends_on:
+#      commerce-db:
+#        condition: service_healthy
+#      configserver:
+#        condition: service_started
+#    ports:
+#      - "8081:8081"
+#    restart: on-failure:5
+
+  customer-service:
+    image: ecommerce/customer-service:0.0.1-SNAPSHOT
+    environment:
+      SPRING_PROFILES_ACTIVE: "dev"
+      SPRING_CLOUD_CONFIG_URI: "http://configserver:8089"
+      CONFIGSERVER_PORT: "8089"
+      DATABASESERVER_PORT: "5433"
+      ENCRYPT_KEY: "IMSYMMETRIC"
+    depends_on:
+      customer-db:
+        condition: service_healthy
+      configserver:
+        condition: service_started
+    ports:
+      - "8080:8080"
+    restart: on-failure:5
