+++ configserver/src/main/java/com/ecommerce/configserver/ConfigurationServerApplication.java
@@ -1,4 +1,4 @@
-package com.ecommerce;
+package com.ecommerce.configserver;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
