+++ configserver/src/main/resources/config/commerce-service-dev.properties
@@ -1,4 +1,4 @@
-example.property= I AM THE DEFAULT
+example.property= I AM DEV
 
 
 spring.datasource.url=jdbc:postgresql://commerce-db:5432/commerce_db
