+++ configserver/src/main/resources/config/customer-service-dev.properties
@@ -1,4 +1,4 @@
-example.property= I AM THE DEFAULT
+example.property= I AM DEV
 
 
 spring.datasource.url=jdbc:postgresql://customer-db:5432/consumer_db
