+++ configserver/src/test/java/com/ecommerce/configserver/ConfigurationServiceApplicationTests.java
@@ -1,10 +1,10 @@
-package com.ecommerce;
+package com.ecommerce.configserver;
+
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ConfigurationServerApplicationTests {
-
+public class ConfigurationServiceApplicationTests {
     @Test
     void contextLoads() {
     }
