+++ customer-service/src/main/java/com/ecommerce/customer/CustomerServiceApplication.java
@@ -1,4 +1,4 @@
-package com.ecommerce;
+package com.ecommerce.customer;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.cloud.context.config.annotation.RefreshScope;
 
 @SpringBootApplication
 @RefreshScope
-public class CommerceServiceApplication {
+public class CustomerServiceApplication {
     public static void main(String[] args) {
-        SpringApplication.run(CommerceServiceApplication.class, args);
+        SpringApplication.run(CustomerServiceApplication.class, args);
     }
 }
