--- customer-service/src/test/java/com/ecommerce/CustomerServiceApplicationTests.java
@@ -1,11 +1,0 @@
-package com.ecommerce;
-
-import org.junit.jupiter.api.Disabled;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CustomerServiceApplicationTests {
-    @Disabled
-    void contextLoads() {
-    }
-}
